Combine previously introduced types API with current new more comprehensive set.
