+++ python/cudf/cudf/api/types.py
@@ -4,6 +4,7 @@
 from collections.abc import Sequence
 from functools import wraps
 from inspect import isclass
+from typing import List, Union
 
 import cupy as cp
 import numpy as np
@@ -317,7 +318,41 @@ def _wrap_pandas_is_dtype_api(func):
     return wrapped_func
 
 
-union_categoricals = pd_types.union_categoricals
+def _union_categoricals(
+    to_union: List[Union[cudf.Series, cudf.Index]],
+    sort_categories: bool = False,
+    ignore_order: bool = False,
+):
+    """
+    This is an internal API which combines categorical data.
+    """
+    # TODO(s) in the order specified :
+    # 1. The return type needs to be changed
+    #    to cudf.Categorical once it is implemented.
+    # 2. Make this API public (i.e., to resemble
+    #    pd.api.types.union_categoricals)
+
+    if ignore_order:
+        raise TypeError("ignore_order is not yet implemented")
+
+    result_col = cudf.core.column.CategoricalColumn._concat(
+        [obj._column for obj in to_union]
+    )
+    if sort_categories:
+        sorted_categories = result_col.categories.sort_by_values(
+            ascending=True
+        )[0]
+        result_col = result_col.cat().reorder_categories(
+            new_categories=sorted_categories
+        )
+
+    return cudf.Index(result_col)
+
+
+# TODO: The below alias is removed for now since improving cudf categorical
+# support is ongoing and we don't want to introduce any ambiguities. The above
+# method _union_categoricals will take its place once exposed.
+# union_categoricals = pd_types.union_categoricals
 infer_dtype = pd_types.infer_dtype
 pandas_dtype = pd_types.pandas_dtype
 is_bool_dtype = pd_types.is_bool_dtype
