--- python/cudf/cudf/api/types/__init__.py
@@ -1,3 +1,0 @@
-# Copyright (c) 2021, NVIDIA CORPORATION.
-
-from cudf.api.types.categoricals import _union_categoricals
